refactor: extracted map to a component
