+++ web-client/src/pages/MapManagement/ManageMapPage.tsx
@@ -1,12 +1,4 @@
-import {
-  GoogleMap,
-  Marker,
-  MarkerF,
-  useJsApiLoader,
-} from "@react-google-maps/api";
-import { GOOGLE_MAP_API_KEY } from "../../api.config";
-// import { lightMapTheme } from "./mapStyle";
-import { useState } from "react";
+import { MarkerF } from "@react-google-maps/api";
 import FacilityForm from "./components/FacilityForm";
 import FacilitiesList from "./components/FacilitiesList";
 import { useGetFacilitiesQuery } from "../../services/facilityQuery";
@@ -18,26 +10,13 @@ import {
   setAddMode,
   selectAddMode,
   selectionFacility,
-  setTempMarkerPos,
   selectTempMarkerPos,
 } from "../../store/slices/facilitySlice";
-
-const containerStyle = {
-  width: "100%",
-  height: "100vh",
-};
-// ,
-const center = {
-  lat: 14.860767193574064,
-  lng: 120.81013409214616,
-};
+import MapComponent from "./components/MapComponent";
+import { useState } from "react";
 
 const ManageMapPage = () => {
   const [map, setMap] = useState<google.maps.Map | null>(null);
-  const { isLoaded: isMapLoaded } = useJsApiLoader({
-    id: "google-map-script",
-    googleMapsApiKey: GOOGLE_MAP_API_KEY,
-  });
 
   const dispatch = useAppDispatch();
   const selectedFacility = useAppSelector(selectionFacility);
@@ -54,12 +33,9 @@ const ManageMapPage = () => {
   if (isFacilitiesFetchError) {
     return <p>Something went wrong...</p>;
   }
-
-  const onMapClickHandler = (event: google.maps.MapMouseEvent) => {
-    if (!addMode || !map) return;
-    dispatch(
-      setTempMarkerPos({ lat: event.latLng?.lat(), lng: event.latLng?.lng() })
-    );
+  // Set map state
+  const onMapLoad = (map: google.maps.Map) => {
+    setMap(map);
   };
 
   const panMapTo = (lat: number, lng: number) => {
@@ -74,7 +50,6 @@ const ManageMapPage = () => {
 
   return (
     <div className="relative h-screen">
-      <h1>Manage Map Page</h1>
       {isFacilitiesLoading && <p>Loading map details...</p>}
       <div className="relative z-10 flex flex-row gap-2 w-max">
         <div className="flex flex-col z-10 bg-white p-2 gap-2">
@@ -110,57 +85,28 @@ const ManageMapPage = () => {
         {/* IF SELECTED FACILITY, show facility form */}
         {selectedFacility && <FacilityForm facility={selectedFacility} />}
       </div>
-      {isMapLoaded && (
-        <div className="absolute top-0 z-0 w-full">
-          <GoogleMap
-            mapContainerStyle={containerStyle}
-            center={center}
-            zoom={15}
-            options={{
-              mapId: "ca99ebef66d0dc2e",
-              // styles: lightMapTheme,
-              minZoom: 13,
-              maxZoom: 16,
-              zoomControl: false,
-              streetViewControl: false,
-              fullscreenControl: false,
-              mapTypeControl: false,
-            }}
-            onLoad={(map) => {
-              setMap(map);
-            }}
-            onClick={(event) => onMapClickHandler(event)}
-          >
-            {/* Child components, such as markers, info windows, etc. */}
-            {!isFacilitiesLoading &&
-              facilities?.map((facility) => (
-                <MarkerF
-                  key={facility._id}
-                  position={{
+      <div className="absolute top-0 z-0 w-full">
+        <MapComponent onSetMapHandler={onMapLoad}>
+          {/* Child components, such as markers, info windows, etc. */}
+          {!isFacilitiesLoading &&
+            facilities?.map((facility) => (
+              <MarkerF
+                key={facility._id}
+                position={{
+                  lat: facility.latitude,
+                  lng: facility.longitude,
+                }}
+                onClick={() => {
+                  dispatch(setSelectedFacility(facility));
+                  map?.panTo({
                     lat: facility.latitude,
                     lng: facility.longitude,
-                  }}
-                  onClick={() => {
-                    dispatch(setSelectedFacility(facility));
-                    map?.panTo({
-                      lat: facility.latitude,
-                      lng: facility.longitude,
-                    });
-                  }}
-                />
-              ))}
-            {tempMarkerPos && addMode && (
-              <Marker
-                key={"Hello"}
-                position={{
-                  lat: tempMarkerPos.lat ?? 0,
-                  lng: tempMarkerPos.lng ?? 0,
+                  });
                 }}
               />
-            )}
-          </GoogleMap>
-        </div>
-      )}
+            ))}
+        </MapComponent>
+      </div>
     </div>
   );
 };
